Minor fixes to polish model-setup.md
Add missing apostroph in text.

In the code block to check for model availability, make only one empty line between using statement and code. This is consistent with all the other parts in the Windows AI documentation (Except for a single codeblock in the Phi Silica overview :), where I'll create another PR for).